requirements.txt
# YouTube 콘텐츠 분석 시스템 - 필수 라이브러리

# 기본 라이브러리
pandas>=1.3.5
numpy>=1.20.0
matplotlib>=3.5.1
python-dotenv>=0.20.0

# API 클라이언트
google-api-python-client>=2.70.0
google-auth>=2.16.0
google-auth-oauthlib>=0.5.3
google-auth-httplib2>=0.1.0

# YouTube 스크립트 추출
youtube-transcript-api>=0.5.0

# HTTP 요청
requests>=2.28.0

# NLP 관련
nltk>=3.7
spacy>=3.4.0
yake>=0.4.8
rake-nltk>=1.0.6
keybert>=0.7.0
wordcloud>=1.8.2

# 한국어 처리 (선택)
konlpy>=0.6.0

# Streamlit 웹 인터페이스
streamlit>=1.17.0

# (선택) Google Ads API - Google Keyword Planner 사용 시
# google-ads>=19.0.0 